Rename defaultMessages to httpCodeMessages
